This change extracts socket logic into a reusable `useSocket` composable, simplifying the Dashboard component by removing socket connection and event handling code and replacing it with calls to the composable's provided functions and reactive state.
